+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Updated the right permissions for the `install-boxlang`
+
 ## [1.7.0] - 2025-06-05
 
 ### Added